refactor(editor-template-plugins): include modal and button in MetaFields component
